==> 1.download_data.ipynb <==
import pandas as pd
import wget

metadata = pd.read_csv("/data/hoan/amromics/prediction/data/Ecoli1936metafiles/metadata.csv")

metadata.head()

assembly_file = pd.read_csv("/data/hoan/amromics/prediction/data/Ecoli1936metafiles/assembly_file.tsv", sep='\t')

assembly_file.head()

metadata.iloc[2,1]

count = 0
for isolate in metadata.iloc[:,1]:
    # print(isolate, end=',')
    # assembly_file.loc[assembly_file['sample_title'] == isolate]['submitted_ftp'].values
    if isolate in assembly_file['sample_title'].values:
        # print(isolate)
        url = 'https://'+assembly_file.loc[assembly_file['sample_title'] == isolate]['submitted_ftp'].values[0]
        count = count + 1
        # print(url)
        wget.download(url, out='/data/hoan/amromics/amr_prediction/data/assembly')

print(count)

isolate = []
sample = []
for i in range(len(assembly_file.index)):
    if assembly_file.iloc[i, 6] in metadata.iloc[:,1].values:
        isolate.append(assembly_file.iloc[i, 6])
        sample.append(assembly_file.iloc[i, 2])


isolate[:10]

sample[:10]

df = pd.DataFrame({'sample_accession': sample, 'isolate': isolate})

df.head()

df.to_csv('/data/hoan/amromics/prediction/data/Ecoli1936metafiles/sample_isolate.csv',index=False)

newdf = pd.read_csv('/data/hoan/amromics/prediction/data/Ecoli1936metafiles/sample_isolate.csv')

newdf
