==> 2.runProkka-revision.ipynb <==
import os
import traceback
import logging
import os.path
import networkx as nx
import pandas as pd
import numpy as np
import pylab as pl
from matplotlib import collections  as mc
import glob
import re
import csv
import shutil

# try:
#     %load_ext autoreload
#     %autoreload 2
# except Exception as e:
#     logging.error(traceback.format_exc())

# simversion = '_v01'
simversion = '_plasmid_v01'
simversionPangraph = '_assemBiGraph2_v01' 
# 1 True, 0 False
run_prokka = 1
art_simulator = 0
run_art = 0
run_spades = 0
run_shovill = 0
run_panta = 1
split_paralogs = 1
run_multicsar = 0
run_ragout = 0
run_pangraph = 1
pangenome_data = '/data/hoan/amromics/prediction/data/Ecoli1936/'

if 'Kp' in pangenome_data:
    data_base = ' --genus Klebsiella --species pneumoniae --cpus 30'
elif 'Ecoli' in pangenome_data:
    data_base = ' --genus Escherichia --species coli --cpus 30'
else:
    data_base = ' --cpus 30'
print("Use the correct database for the species: ", data_base)
print("Use the correct database for the species: ", data_base)

output_dir = pangenome_data + 'temp'
# prokka_dir = pangenome_data + 'prokka'
prokka_dir = pangenome_data + 'prokkaevalue1minus4'
os.system('mkdir '+ prokka_dir)
conda_dir = 'source ~/miniconda3/etc/profile.d/conda.sh && conda activate amromics && '
if run_prokka:
    fffile = pangenome_data + '*.fa'
    list_database = glob.glob(fffile)
    for i in range(len(list_database)):
        input_data = list_database[i]
        print(input_data)
        file_name = input_data.split("/")[-1][:-4]
        prokka_bin = conda_dir + 'prokka --outdir ' + output_dir +data_base +' --prefix ' + file_name +' '+input_data
        os.system(prokka_bin)
        os.system('cp ' + output_dir + '/'+file_name+'.gff ' + prokka_dir)
        os.system('rm -r '+ output_dir)
        # if i>=1:
        #     break;

### Then, run panta

# Instruction to run panta here: https://github.com/amromics/panta

# Run Panta in: /data/hoan/amromics/pantaPred/ file runme.sh (env amromics)

# panta main -o /data/hoan/amromics/prediction/output/pantaEcoli1936align_v4 -g /data/hoan/amromics/prediction/data/Ecoli1936/prokka/*.gff -a protein
# panta main -o /data/hoan/amromics/prediction/output/pantaEcoli1936align_v5 -g /data/hoan/amromics/prediction/data/Ecoli1936/prokka/*.gff -a protein -i 0.95
# panta main -o /data/hoan/amromics/prediction/output/pantaEcoli1936align_v6 -g /data/hoan/amromics/prediction/data/Ecoli1936/prokka/*.gff -s -i 0.95

### Prepare data for panta -add

metadata_panta = pd.read_csv("revision/metadataEcoliST.csv")
metadata_panta.head(2)

sample_isolate = pd.read_csv('/data/hoan/amromics/prediction/data/Ecoli1936metafiles/sample_isolate.csv')
sample_isolate.head(2)
isolate2sample = {}
for idx in range(len(sample_isolate.index)):
    isolate2sample[sample_isolate.iloc[idx,1]] =  sample_isolate.iloc[idx,0]+'.contig.gff'

isolate2sample['11658_4#1']

# test1_idx = [131, 95]
# train_dir = '/data/hoan/amromics/prediction/data/Ecoli1936/prokka_train/train1'
# test_dir = '/data/hoan/amromics/prediction/data/Ecoli1936/prokka_test/test1'
test1_idx = [73, 10, 405, 38, 393]
train_dir = '/data/hoan/amromics/prediction/data/Ecoli1936/prokka_train/train2'
test_dir = '/data/hoan/amromics/prediction/data/Ecoli1936/prokka_test/test2'
if not os.path.isdir(train_dir):
    os.system('mkdir '+ train_dir)
if not os.path.isdir(test_dir):
    os.system('mkdir '+ test_dir)

for i in range(len(metadata_panta.index)):
    filedir = '/data/hoan/amromics/prediction/data/Ecoli1936/prokka/' + isolate2sample[metadata_panta.iloc[i, 0]]
    cmd1 = 'cp '+ filedir + ' ' + train_dir
    cmd2 = 'cp '+ filedir + ' ' + test_dir
    if metadata_panta.iloc[i, -1] in test1_idx:
        os.system(cmd2)
    else:
        os.system(cmd1)

len(metadata_panta.index)
