==> 3.b.preprocessDataPanta.ipynb <==
## Idea: Using length + coverage in the scoring function
# cluster va annoated_cluster ko phai tuong ung 1-1

import os
import traceback
import logging
import os.path
import networkx as nx
import pandas as pd
import numpy as np
import pylab as pl
from matplotlib import collections  as mc
import glob
import re
import csv
import shutil
from sklearn import preprocessing
import json
import gzip
from Bio import SeqIO
from Bio.Seq import Seq
from sklearn import preprocessing
from sklearn.feature_selection import VarianceThreshold

# from pangraph.utils import overlap
# overlap("AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA1AAAAA", "AAAAAA1BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB")

# simversion = '_v01'
simversion = '_plasmid_v01'
pangenome_data = '/data/hoan/amromics/prediction/output/pantaEcoli1936align_v4/'
scaffold_out_dir = '/data/hoan/amromics/genome-graph/scaffold_output/'

try:
    %load_ext autoreload
    %autoreload 2
except Exception as e:
    logging.error(traceback.format_exc())

from pangraph.utils import export_metadata

## Export file for Pasa
panta_output = '/data/hoan/amromics/prediction/output/pantaEcoli1936align_v4/'
# export_metadata(panta_output)

### Run pangraph

data_dir = panta_output
sample_info = pd.read_csv(data_dir + "/samples.tsv", delimiter='\t', header=None)
sample_info.columns = ['Name', 'SampleID']
gene_info = pd.read_csv(data_dir + "/gene_info.tsv", delimiter='\t', header=None)
gene_info.columns =['GeneName', 'SampleID', 'clusterID']
gene_position = pd.read_csv(data_dir + '/gene_position.tsv', delimiter='\t', header=None)
gene_position.columns =['SampleID', 'ContigName', 'GeneSequence']
# sort by length of contigs
gene_position.sort_values(by="GeneSequence", key=lambda x: x.str.len(),  ascending=False, inplace=True)
n_samples = len(np.unique(gene_position.iloc[:,0]))

try:
    %load_ext autoreload
    %autoreload 2
except Exception as e:
    logging.error(traceback.format_exc())
from pangraph import PanGraph

pangraph = PanGraph(sample_info, gene_info, gene_position)

H = pangraph.construct_simple_graph(method = "graph_alignment", sample_id_ref = None,  min_nucleotides = 10, min_genes = 0, edge_weight="unit")

H.number_of_nodes()

list(H.nodes())[:3]

dict_clusters=json.load(open(os.path.join(panta_output,'clusters.json')))
# map_geneid_to_numberic_cid={}
# cindex=0
# for k in dict_clusters.keys():
#     # print(k)
#     # print(dict_clusters[k])
#     map_geneid_to_numberic_cid[k] =cindex

#     for g in dict_clusters[k]:
#         map_geneid_to_numberic_cid[g]=cindex
#     cindex=cindex+1

idx2representative = {}
for idx, key in enumerate(dict_clusters):
    idx2representative[idx] = key

ann_clusters=json.load(open(os.path.join(panta_output,'annotated_clusters.json')))

for idx, key in enumerate(ann_clusters):
    print(idx, key)
    if idx > 2:
        break

representative2genecluster = {}
for key in ann_clusters:
    representative2genecluster[ann_clusters[key]['representative']] = key

representative2genecluster[idx2representative[0]]

sorted_node = sorted(H.degree, key=lambda x: x[1], reverse=True)

sorted_node[:10]

topnodes = []
for i in range(400):
    # print (sorted_node[i][0][2:])
    idx = int(sorted_node[i][0][2:])
    if idx2representative[idx] in representative2genecluster:
        # print(idx2representative[idx], representative2genecluster[idx2representative[idx]])
        topnodes.append(representative2genecluster[idx2representative[idx]])
        # Add neighbors to topnodes
        for node in H.neighbors(sorted_node[idx][0]):
            idx = int(node[2:])
            if idx2representative[idx] in representative2genecluster:
                # print(idx2representative[idx], representative2genecluster[idx2representative[idx]])
                topnodes.append(representative2genecluster[idx2representative[idx]])
topnodes = list(set(topnodes))

len(topnodes), topnodes[:3]

### Compute label encoder for AMR cluster

pantaout_dir = panta_output
with open(pantaout_dir + 'samples.json', 'r') as JSON:
    sample_dict = json.load(JSON)
sample2integerindex = {}
for idx in range(len(sample_dict)):
    sample2integerindex[sample_dict[idx]['id']] = idx
n_samples = len(sample_dict)

amr_mat = None;
start_idx = [0];
# end_idx = [];
for idx in range(len(topnodes)):
    alignment_dir = pantaout_dir + 'clusters/' + topnodes[idx] +'/'+topnodes[idx]+'.faa.aln.gz'
    # https://www.biostars.org/p/710/
    # fasta_sequences = SeqIO.parse(open(alignment_dir),'fasta')
    # for fasta in fasta_sequences:
    #     name, sequence = fasta.id, str(fasta.seq)
    #     print(name)
    #     print(sequence)
    codes = ['A', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'K', 'L',
             'M', 'N', 'P', 'Q', 'R', 'S', 'T', 'V', 'W', 'Y', '-']
    le = preprocessing.LabelEncoder()
    le.fit(codes)
    mat = None
    index = 0
    index_set = []
    with gzip.open(alignment_dir, "rt") as handle:
        for record in SeqIO.parse(handle, "fasta"):
            name, sequence = record.id, str(record.seq)
            sample_id = name.split('-')[0]
            # print(name)
            # print(sequence)
            # print(1 + le.transform([*sequence]))
            if index == 0:
                mat = np.zeros((n_samples, len(sequence)))
            index += 1
            mat[sample2integerindex[sample_id],:] = 1 + le.transform([*sequence])
            index_set.append(sample2integerindex[sample_id])
            # print(record.id)
    if idx==0:
        start_idx += [start_idx[-1] + mat.shape[1]]
        amr_mat = mat
    else:
        # ## Run feature selection
        variant_thres = 0.05
        vs = True
        if len(index_set) >= 5:
            try:
                sel = VarianceThreshold(variant_thres)
                sel.fit(mat[index_set,:])
            except ValueError:
                vs = False
            if vs:
                mat = mat[:, sel.variances_>variant_thres]
                if mat.shape[0] > 0:
                    start_idx += [start_idx[-1] + mat.shape[1]]
                    amr_mat = np.append(amr_mat, mat, axis=1)
end_idx = start_idx[1:]
start_idx = start_idx[:-1]

np.save(pantaout_dir + 'pantaGFilterHighGeneNeighborVT5.npy', amr_mat) # save numpy array

amr_mat.shape

amr_mat = None;
start_idx = [0];
# end_idx = [];
for idx in range(len(topnodes)):
    alignment_dir = pantaout_dir + 'clusters/' + topnodes[idx] +'/'+topnodes[idx]+'.faa.aln.gz'
    # https://www.biostars.org/p/710/
    # fasta_sequences = SeqIO.parse(open(alignment_dir),'fasta')
    # for fasta in fasta_sequences:
    #     name, sequence = fasta.id, str(fasta.seq)
    #     print(name)
    #     print(sequence)
    codes = ['A', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'K', 'L',
             'M', 'N', 'P', 'Q', 'R', 'S', 'T', 'V', 'W', 'Y', '-']
    le = preprocessing.LabelEncoder()
    le.fit(codes)
    mat = None
    index = 0
    index_set = []
    with gzip.open(alignment_dir, "rt") as handle:
        for record in SeqIO.parse(handle, "fasta"):
            name, sequence = record.id, str(record.seq)
            sample_id = name.split('-')[0]
            # print(name)
            # print(sequence)
            # print(1 + le.transform([*sequence]))
            if index == 0:
                mat = np.zeros((n_samples, len(sequence)))
            index += 1
            mat[sample2integerindex[sample_id],:] = 1 + le.transform([*sequence])
            index_set.append(sample2integerindex[sample_id])
            # print(record.id)
    start_idx += [start_idx[-1] + mat.shape[1]]
    if idx==0:
        amr_mat = mat
    else:
        amr_mat = np.append(amr_mat, mat, axis=1)
end_idx = start_idx[1:]
start_idx = start_idx[:-1]

np.save(pantaout_dir + 'pantaGFilterHighGeneNeighbor.npy', amr_mat) # save numpy array

# amrgene_annotation = pd.DataFrame({'gene': amr_clusterID, 'start_index': start_idx, 'end_index': end_idx})
amrgene_annotation = pd.DataFrame({'start_index': start_idx, 'end_index': end_idx})
amrgene_annotation.to_csv(pantaout_dir + 'pantaGFilterHighGeneNeighbor_geneindex.csv', index=None)
