==> 7.analyze_features-revision.ipynb <==
import pandas as pd
import seaborn as sns
import glob
from numpy import genfromtxt
# from sklearn.metrics import f1_score
from matplotlib import pyplot as plt
import numpy as np
import json
from collections import OrderedDict
import os
import re
import logging
import multiprocessing
from functools import partial
from datetime import datetime
from Bio import SeqIO
from Bio.Seq import Seq
import gzip
from sklearn import preprocessing
from sklearn.feature_selection import VarianceThreshold
# https://pandas.pydata.org/docs/getting_started/intro_tutorials/03_subset_data.html

### Find SNPs info

pantaout_dir = '/data/hoan/amromics/prediction/output/pantaEcoli1936align_v9/'
with open(pantaout_dir + 'samples.json', 'r') as JSON:
    sample_dict = json.load(JSON)
sample2integerindex = {}
for idx in range(len(sample_dict)):
    sample2integerindex[sample_dict[idx]['id']] = idx
n_samples = len(sample_dict)
amrgene_annotation = pd.read_pickle(pantaout_dir + 'SNPsCoreGeneFullSubmission_metadata.pkl')

amrgene_annotation.head(2) # index in SNPs matrix

snpID = 'yfcR@192641' # 192641 corresponds to the index in the SNPs matrix

def getSNPPositionIndex(snpID):
    ## Compute the SNP matrix
    geneID = snpID.split('@')[0]
    start_index = amrgene_annotation.loc[amrgene_annotation['gene']==geneID].iloc[0,1]
    snpPositionWithinGene = int(snpID.split('@')[1]) - start_index
    alignment_dir = pantaout_dir + 'clusters/' + geneID +'/'+geneID+'.faa.aln.gz'
    codes = ['A', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'K', 'L',
             'M', 'N', 'P', 'Q', 'R', 'S', 'T', 'V', 'W', 'Y', '-', 'X']
    le = preprocessing.LabelEncoder()
    le.fit(codes)
    mat = None; index = 0; index_set = []
    with gzip.open(alignment_dir, "rt") as handle:
        for record in SeqIO.parse(handle, "fasta"):
            name, sequence = record.id, str(record.seq)
            sample_id = name.split('-')[0]
            if index == 0:
                mat = np.zeros((n_samples, len(sequence)))
            index += 1
            mat[sample2integerindex[sample_id],:] = 1 + le.transform([*sequence])
            index_set.append(sample2integerindex[sample_id])

    variant_thres = 0
    vs = True
    try:
        sel = VarianceThreshold(variant_thres)
        sel.fit(mat[index_set,:])
    except ValueError:
        vs = False
    if vs:
        mat = mat[:, sel.variances_>variant_thres]
        
    ## Compute the index
    count = 0
    truePosition = None
    boolVec = sel.variances_>variant_thres
    for idx in range(len(boolVec)):
        if boolVec[idx]:
            count += 1
        if count == snpPositionWithinGene:
            truePosition = count
    positionIndex = geneID +'@'+str(truePosition)
    return(positionIndex)

positionIndex = getSNPPositionIndex(snpID)

positionIndex

### Find Kmer info

kmerID = 'VSLHINQGEI'

from pangraph.utils import parse_gff_AMRgene_finder, parse_gff_AMRgene_finder_BARRGD_ID
with open(pantaout_dir + 'annotated_clusters.json', 'r') as JSON:
    json_dict = json.load(JSON)
gene2clusterdict = {}
for key in json_dict:
    if len(json_dict[key])==0:
        gene2clusterdict[key] = key
    for gene in json_dict[key]['gene_id']:
        gene2clusterdict[gene] = key

amr_gene = []
pantain_dir = '/data/hoan/amromics/prediction/data/Ecoli1936/prokka/'
for data_dir in glob.glob(pantain_dir + '*.gff'):
    # print(data_dir)
    in_fh = open(data_dir)
    sample_id = data_dir.split('/')[-1][:-4]
    amr_gene += parse_gff_AMRgene_finder(in_fh, sample_id)
    in_fh.close()

amr_clusterID = [gene2clusterdict[gene] for gene in amr_gene]
amr_clusterID = list(set(amr_clusterID))

amr_gene[0:1], len(amr_clusterID), amr_clusterID[0:4]

def getAMRPositionIndex(kmerID):
    ksize = 10; # k = 10 for protein, 20 for DNA
    topAMRList = [];
    pairdata = []
    positionIndexVec = []
    computed_gene_cluster = amr_clusterID
    for idx in range(len(computed_gene_cluster)):
        alignment_dir = pantaout_dir + 'clusters/' + computed_gene_cluster[idx] +'/'+computed_gene_cluster[idx]+'.faa.aln.gz'
        # alignment_dir = pantaout_dir + 'clusters/' + computed_gene_cluster[idx] +'/'+computed_gene_cluster[idx]+'.fna.aln.gz'
        clusterBool = False
        with gzip.open(alignment_dir, "rt") as handle:
            for record in SeqIO.parse(handle, "fasta"):
                name, sequence = record.id, str(record.seq)
                sample_id = name.split('-')[0]
                seqraw = sequence.replace('-', '')
                n_kmers = len(seqraw) - ksize + 1
                # print(idx, end =',')
                breakLoop = False
                for i in range(n_kmers):
                    kmer = seqraw[i:i + ksize] 
                    if kmer == kmerID:
                        clusterBool = True
                        pairdata.append((computed_gene_cluster[idx] , sample_id))
                        positionIndexVec.append(computed_gene_cluster[idx]+'@'+str(i)+'@'+str(i+10))
                        # breakLoop = True;
                        # break;
        if clusterBool:
            topAMRList.append(computed_gene_cluster[idx])
    positionIndexVecUnique = np.unique(positionIndexVec)
    positionAMRIndex = ''
    for v in positionIndexVecUnique:
        positionAMRIndex = positionAMRIndex + v + ';'
    positionAMRIndex = positionAMRIndex[:-1]
    return(positionAMRIndex)

positionIndex = getAMRPositionIndex(kmerID)

positionIndex

def getTypeVariant(variantID):
    if '@' in variantID:
        return ('SNP')
    elif '_' in variantID:
        return ('PAGene')
    elif len(variantID) != 10:
        return ('PAGene')
    else:
        return ('AMRKmer')

# Opening JSON file
scoreInputFileDir = 'data/TopFeatures/gain_ft_importance.json'
# 'data/TopFeatures/split_ft_importance.json'
with open(scoreInputFileDir) as json_file:
    gainImportance = json.load(json_file)

# gainImportance

# for v in gainImportance['AMP'][0]:
#     print(v, getTypeVariant(v))

def Reverse(lst):
   new_lst = lst[::-1]
   return new_lst

dfVec = []
nFeatures = 10
count = 0
for antibiotics in gainImportance:
    # count += 1
    # if count == 3:
    #     break
    print('Reverse the list')
    featureList = Reverse(gainImportance[antibiotics][0])
    scoreList = Reverse(gainImportance[antibiotics][1])
    print(antibiotics)
    for i in range(nFeatures):
        print(i, end = ',')
        feature = featureList[i]
        score = scoreList[i]
        featureType = getTypeVariant(feature)
        positionIndex = None
        if featureType == 'SNP':
            positionIndex = getSNPPositionIndex(feature)
        elif featureType == 'AMRKmer':
            positionIndex =  getAMRPositionIndex(feature)
        else:
            positionIndex = feature
        dfVec.append([antibiotics, feature, featureType, i, score, positionIndex])

df = pd.DataFrame(dfVec)
df.columns = ['antibiotics', 'featureName', 'variantType', 'rank', 'featureScore', 'positionIndex']

if 'gain' in scoreInputFileDir:
    df.to_csv('data/TopFeatures/gain_feature_annotation.csv', index=False)
else:
    df.to_csv('data/TopFeatures/split_feature_annotation.csv', index=False)

### Copy clusters to directory

df_file = pd.read_csv('data/TopFeatures/gain_feature_annotation.csv')

df_file.head(2)

# for i in range(len(df_file.index)):
#     if (df_file.iloc[i,2]=='AMRKmer') or (df_file.iloc[i,2]=='SNP'):
#         positionIndex = df_file.iloc[i,5].split(';')
#         for cluster in positionIndex:
#             gene = cluster.split('@')[0]
#             print(gene, end = ',')
#             alignment_dir = pantaout_dir + 'clusters/' + gene +'/'+gene+'.faa.aln.gz'
#             os.system('cp '+ alignment_dir + ' data/TopFeatures/featureClusters')

gene_annotation = pd.read_csv(pantaout_dir + 'gene_annotation.csv')

# gene_annotation

amr_gene = []; AMR_anno = [];
pantain_dir = '/data/hoan/amromics/prediction/data/Ecoli1936/prokka/'
for data_dir in glob.glob(pantain_dir + '*.gff'):
    # print(data_dir)
    in_fh = open(data_dir)
    sample_id = data_dir.split('/')[-1][:-4]
    amr_gene_sample, AMR_anno_sample= parse_gff_AMRgene_finder_BARRGD_ID(in_fh, sample_id)
    amr_gene += amr_gene_sample
    AMR_anno += AMR_anno_sample
    in_fh.close()

geneID2NCBI = {}
for gene, itemID in AMR_anno:
    geneID2NCBI[gene] = itemID

clusterID = 'emrB_1'
NCBIList = []
for gene in json_dict[clusterID]['gene_id']:
    if gene in geneID2NCBI:
        print(gene2clusterdict[gene], geneID2NCBI[gene])
        NCBIList.append(geneID2NCBI[gene])

from collections import Counter

def most_frequent_element(lst):
    counts = Counter(lst)
    most_common = counts.most_common(1)
    return most_common[0][0]

# Example usage:
print(most_frequent_element(NCBIList))  # Output: 3

def get_NCBIID(clusterID):
    NCBIList = []
    for gene in json_dict[clusterID]['gene_id']:
        if gene in geneID2NCBI:
            NCBIList.append(geneID2NCBI[gene])
    if len(NCBIList) > 0:
        return(most_frequent_element(NCBIList))
    else:
        return ('')

get_NCBIID(clusterID)

np.unique(gene_annotation.loc[gene_annotation['gene_name'] == clusterID]['gene_product'].values) # many gene same with the same cluster

most_frequent_element(gene_annotation.loc[gene_annotation['gene_name'] == clusterID]['gene_product'].values)

# most_frequent_element([1,1,2,2,2])

df_file.head(2)

# antibiotics, featureID (Name), variantType, rank, featureScore, Gene location (position Index), Gene, Gene Function, NCBI ID

df_full = []
for i in range(len(df_file.index)):
    positionIndex = df_file.iloc[i, 5].split("@")
    print(i,end=',')
    if len(positionIndex) == 3:
        gene = positionIndex[0]
        position = positionIndex[1]+':'+positionIndex[2]
    elif len(positionIndex) == 2:
        gene = positionIndex[0]
        position = positionIndex[1]
    else:
        gene = positionIndex[0]
        position = ''
    NCBIID = get_NCBIID(gene)
    if gene in gene_annotation['gene_name'].values:
        function = most_frequent_element(gene_annotation.loc[gene_annotation['gene_name'] == gene]['gene_product'].values)
    else:
        function = ''
    if gene in amr_clusterID:
        AMRCluster = 'Yes'
    else:
        AMRCluster = 'No'
    row = [df_file.iloc[i, 0], df_file.iloc[i, 1], df_file.iloc[i, 2], int(df_file.iloc[i, 3]) + 1, df_file.iloc[i, 4], position,gene,function,NCBIID, AMRCluster]
    df_full.append(row)

df = pd.DataFrame(df_full)
df.columns = ['Antibiotics', 'Feature ID', 'Variant Type', 'Rank', 'Feature Score', 'Gene Location', 'Gene', 'Gene Function', 'NCBI ID', 'AMR Gene']

df.to_csv('revision/feature_annotation.csv', index=None)

gene in gene_annotation['gene_name']

# amr_clusterID
